% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.density.R
\name{as.data.frame.density}
\alias{as.data.frame.density}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{density}(x, ...)
}
\arguments{
\item{x}{any R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Functions to check if an object is a data frame, or coerce it if possible.
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
