% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor_savagedickey.R
\name{as.numeric.bayesfactor_savagedickey}
\alias{as.numeric.bayesfactor_savagedickey}
\title{Convert to Numeric}
\usage{
\method{as.numeric}{bayesfactor_savagedickey}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Convert to Numeric
}
