% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_map.R
\name{as.numeric.p_map}
\alias{as.numeric.p_map}
\title{Numeric Vectors}
\usage{
\method{as.numeric}{p_map}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Numeric Vectors
}
