% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_significance.R
\name{as.numeric.p_significance}
\alias{as.numeric.p_significance}
\title{Convert to Numeric}
\usage{
\method{as.numeric}{p_significance}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Convert to Numeric
}
