% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{as_numeric_rope}
\alias{as_numeric_rope}
\title{TEMPORARY FUNCTION}
\usage{
as_numeric_rope(x, ...)
}
\arguments{
\item{x}{need to be documented}

\item{...}{need to be documented}
}
\description{
TEMPORARY FUNCTION
}
