% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_p.R
\name{bayes_p}
\alias{bayes_p}
\title{Bayesian p-value}
\usage{
bayes_p(posterior, precision = 2^10)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{precision}{number of points for density estimation. See the `n` parameter in \link[=density]{density}.}
}
\description{
Compute a Bayesian equivalent p value, related to the odds that a parameter (described by its posterior distribution) has againt the null hypothesis (h0) using Mills' (2014, 2017) Objective Bayesian Hypothesis Testing paradigm.
}
\details{
It correponds to the density value at 0 divided by the density of the highest density point.
}
\examples{
library(bayestestR)

bayes_p(posterior = rnorm(1000, 0, 1))
bayes_p(posterior = rnorm(1000, 10, 1))

}
\references{
\href{https://www.youtube.com/watch?v=Ip8Ci5KUVRc}{Mill's talk}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
