% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor_restrict.R
\name{bayesfactor_restrict}
\alias{bayesfactor_restrict}
\alias{bayesfactor_restrict.stanreg}
\alias{bayesfactor_restrict.brmsfit}
\alias{bayesfactor_restrict.emmGrid}
\title{Savage-Dickey density ratio Bayes Factor (BF)}
\usage{
bayesfactor_restrict(posterior, prior = NULL, hypothesis,
  verbose = TRUE, ...)

\method{bayesfactor_restrict}{stanreg}(posterior, prior = NULL,
  hypothesis, verbose = TRUE, effects = c("fixed", "random", "all"),
  ...)

\method{bayesfactor_restrict}{brmsfit}(posterior, prior = NULL,
  hypothesis, verbose = TRUE, effects = c("fixed", "random", "all"),
  ...)

\method{bayesfactor_restrict}{emmGrid}(posterior, prior = NULL,
  hypothesis, verbose = TRUE, ...)
}
\arguments{
\item{posterior}{A \code{stanreg} / \code{brmsfit} object, \code{emmGrid} or a data frame - representing a posterior distribution(s) from (see Details).}

\item{prior}{An object representing a prior distribution (see Details).}

\item{hypothesis}{String ...}

\item{verbose}{Toggle off warnings.}

\item{...}{Currently not used.}

\item{effects}{Should results for fixed effects, random effects or both be returned?
Only applies to mixed models. May be abbreviated.}
}
\value{
A data frame containing the Bayes factor representing evidence \emph{against} the un-restricted model.
}
\description{
This method computes the
}
\details{
Order restriction are conducted by setting an order restriction on the prior and
posterior distributions (\cite{Morey & Wagenmakers, 2013}).
}
\examples{
library(bayestestR)
posterior <- data.frame(X = rnorm(100),
                        X1 = rnorm(100),
                        X3 = rnorm(100))

prior <- data.frame(X = rnorm(100,.4),
                    X1 = rnorm(100,-.2),
                    X3 = rnorm(100))

hyps = c("X > X1 & X1 > X3",
         "X > X1")

bayesfactor_restrict(posterior,prior,hypothesis = hyps)
\dontrun{
# rstanarm models
# ---------------
library(rstanarm)
fit_stan <- stan_glm(mpg ~ wt + cyl + am,
                     data = mtcars)
hyps <- c("am > 0 & cyl < 0","cyl < 0",
          "wt - cyl > 0")
bayesfactor_restrict(fit_stan, hypothesis = hyps)

# emmGrid objects
# ---------------
library(emmeans)
fit <- stan_glm(mpg  ~ factor(cyl), mtcars, family = gaussian())
em_cyl <- emmeans(fit, ~ cyl)
em_cyl
hyps <- c("`4` > 0", # note the ticks!
          "`4` > `6`",
          "`4` > `6` & `6` > `8`")
bayesfactor_restrict(em_cyl, prior = fit, hypothesis = hyps)
}

}
\references{
Morey, R. D., & Wagenmakers, E. J. (2014). Simple relation between Bayesian order-restricted and point-null hypothesis tests. Statistics & Probability Letters, 92, 121-124.
}
