% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_to_prior.R
\name{.extract_priors_rstanarm}
\alias{.extract_priors_rstanarm}
\title{Extract and Returns the priors formatted for rstanarm}
\usage{
.extract_priors_rstanarm(model, ...)
}
\description{
Extract and Returns the priors formatted for rstanarm
}
\keyword{internal}
