% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_to_prior.R
\name{.prior_new_location}
\alias{.prior_new_location}
\title{Set a new location for a prior}
\usage{
.prior_new_location(prior, sign, magnitude = 10)
}
\description{
Set a new location for a prior
}
\keyword{internal}
