% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_samples.R
\name{effective_samples}
\alias{effective_samples}
\alias{effective_samples.brmsfit}
\alias{effective_samples.stanreg}
\title{Effective Sample}
\usage{
effective_samples(model, ...)

\method{effective_samples}{brmsfit}(model, effects = c("fixed", "random",
  "all"), component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL, ...)

\method{effective_samples}{stanreg}(model, effects = c("fixed", "random",
  "all"), parameters = NULL, ...)
}
\arguments{
\item{model}{A \code{stanreg}, \code{stanfit}, or \code{brmsfit} object.}

\item{...}{Not used}

\item{effects}{Should results for fixed effects, random effects or both be returned?
Only applies to mixed models. May be abbreviated.}

\item{component}{Should results for all parameters, parameters for the conditional model
or the zero-inflated part of the model be returned? May be abbreviated. Only
applies to \pkg{brms}-models.}

\item{parameters}{Regular expression pattern that describes the parameters that
should be returned. Meta-parameters (like \code{lp__} or \code{prior_}) are
filtered by default, so only parameters that typically appear in the
\code{summary()} are returned. Use \code{parameters} to select specific parameters
for the output.}
}
\value{
A data frame with two columns: Parameter name and effective sample size.
}
\description{
\code{effective_samples()} returns the effective sample size.
}
\details{
The effective sample size divides the actual sample size by the amount
 of autocorrelation. The effective sample size is a measure of \dQuote{how
 much independent information there is in autocorrelated chains}, or:
 \dQuote{What would be the sample size of a completely non-autocorrelated chain
 that yielded the same information?} (\emph{Kruschke 2015, p182-3}).
 The ratio of effective number of samples and total number of samples
 ranges from 0 to 1, and should be close to 1. The closer this ratio comes
 to zero means that the chains may be inefficient, but possibly still okay.
}
\examples{
\dontrun{
if (require("rstanarm")) {
  m <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
  effective_samples(m)
}}
}
\references{
Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
}
