% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bf.R
\name{format_bf}
\alias{format_bf}
\title{Bayes Factor Formatting}
\usage{
format_bf(bf, stars = FALSE, stars_only = FALSE, name = "BF")
}
\arguments{
\item{bf}{Bayes Factor.}

\item{stars}{Add significance stars (e.g., p < .001***).}

\item{stars_only}{Return only significance stars.}

\item{name}{Name prefixing the text. Can be \code{NULL}.}
}
\value{
A formatted string.
}
\description{
Numeric Values Formatting
}
\examples{
format_bf(1.20)
format_bf(c(1.20, 1557, 3.5, 12), stars = TRUE)
format_bf(c(1.20, 1557, 3.5, 12), name = NULL)
}
