% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_h0.R
\name{odds_h0}
\alias{odds_h0}
\title{Odds againt the null hypothesis (h0)}
\usage{
odds_h0(posterior, precision = 2^10)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{precision}{number of points for density estimation. See the `n` parameter in \link[=density]{density}.}
}
\description{
Compute the odds that a parameter (described by its posterior distribution) has againt the null hypothesis (h0) using Mills' (2014, 2017) Objective Bayesian Hypothesis Testing paradigm.
}
\examples{
library(bayestestR)

posterior <- rnorm(1000, 1, 1)
odds_h0(posterior)

}
\references{
\href{https://www.youtube.com/watch?v=Ip8Ci5KUVRc}{Mill's talk}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
