% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Overlap Coefficient of Two Empirical Distributions}
\usage{
overlap(x, y, precision = 2^10, normalize = FALSE, onesided = FALSE)
}
\arguments{
\item{x}{vector of values from a probability distribution.}

\item{y}{vector of values from a probability distribution.}

\item{precision}{number of points for density estimation. See the `n` parameter in \link[=density]{density}.}

\item{normalize}{use normalized density kernels.}

\item{onesided}{if false (default), it will compute the proportion of the total density mass common to the 2 distributions. Otherwise, it will compute the proportion of the density mass of x "included" in y.}
}
\description{
Calculate the overlap coefficient between two kernel density estimates.
}
\examples{
library(bayestestR)

overlap(x=rnorm(1000, 0, 1), y=rnorm(1000, 0, 1))
overlap(x=rnorm(1000, 0, 1), y=rnorm(1000, 10, 1))

overlap(x=rnorm(1000, 0, 0.1), y=rnorm(1000, 0, 1), normalize=TRUE, onesided=TRUE)
overlap(x=rnorm(1000, 0, 1), y=rnorm(1000, 0, 0.1), normalize=TRUE, onesided=TRUE)

}
\author{
S. Venne
}
