% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope_bounds}
\alias{rope_bounds}
\title{Find Default Equivalence (ROPE) Region Bounds}
\usage{
rope_bounds(model)
}
\arguments{
\item{model}{A Bayesian model.}
}
\description{
Kruschke (2018) suggests that such null value could be set, by default, to the -0.1 to 0.1 range of a standardized parameter (negligible effect size according to Cohen, 1988).
}
\examples{
\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(vs ~ mpg, data = mtcars, family="binomial")
rope_bounds(model)

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
rope_bounds(model)
}

}
