% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope_overlap}
\alias{rope_overlap}
\title{Region of Practical Equivalence based on overlap coefficient (ROPE O)}
\usage{
rope_overlap(posterior, rope = c(-0.1, 0.1))
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{rope}{ROPE's lower and higher bounds.}
}
\description{
Compute how much of the posterior distribution is within a distributed region of practical equivalence.
}
\details{
Variant of traditional \link[=rope]{ROPE}. Instead of the linear proportion of the 95\% CI that falls within the ROPE (assuming that all values of this range have equal likelihood), it is the empirical overlap (with normalized density kernels) between the whole posterior distribution and a normal distribution of mean 0 and SD `range(ROPE)/6`.
}
\examples{
library(bayestestR)

rope_overlap(posterior = rnorm(1000, 0, 0.01), rope = c(-0.1, 0.1))
rope_overlap(posterior = rnorm(1000, 0, 1), rope = c(-0.1, 0.1))
rope_overlap(posterior = rnorm(1000, 1, 0.01), rope = c(-0.1, 0.1))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
