% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope_test}
\alias{rope_test}
\title{Test for Practical Equivalence}
\usage{
rope_test(posterior, bounds = c(-0.1, 0.1), CI = 90, verbose = TRUE)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{bounds}{ROPE's lower and higher bounds.}

\item{CI}{the credible interval to use.}

\item{verbose}{toggle off warnings.}
}
\description{
Perform a Test for Practical Equivalence based on the "HDI+ROPE decision rule" (Kruschke 2018) to check whether parameter values should be accepted or rejected against an explicitely formulated "null hypothesis".
}
\details{
Using the \link[=rope]{ROPE} and the \link[=hdi]{HDI}, Kruschke (2010, 2011, 2014, 2018) suggest using the percentage of the 95\% (or 90\%, considered more stable) \link[=hdi]{HDI} that falls within the ROPE as a decision rule. If the HDI is completely outside the ROPE, the "null hypothesis" for this parameter is "rejected". If the ROPE completely covers the HDI, i.e. all most credible values of a parameter are inside the region of practical equivalence, the null hypothesis is accepted. Else, it’s undecided whether to accept or reject the null hypothesis.
}
\examples{
library(bayestestR)

rope_test(posterior = rnorm(1000, 0, 0.01), bounds = c(-0.1, 0.1))
rope_test(posterior = rnorm(1000, 0, 1), bounds = c(-0.1, 0.1))
rope_test(posterior = rnorm(1000, 1, 0.01), bounds = c(-0.1, 0.1))
rope_test(posterior = rnorm(1000, 1, 1), CI = c(50, 99))

}
\references{
\href{https://strengejacke.wordpress.com/2018/06/06/r-functions-for-bayesian-model-statistics-and-summaries-rstats-stan-brms/}{sjstats}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
