% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor_models.R
\name{update.BFGrid}
\alias{update.BFGrid}
\title{Update BFGrid}
\usage{
\method{update}{BFGrid}(object, subset = NULL, reference = NULL, ...)
}
\arguments{
\item{object}{A BFGrid object.}

\item{subset}{Vector of model indices to keep or remove.}

\item{reference}{Index of model to rereference to, or \code{"top"} to reference to the best model, or \code{"bottom"} to reference to the worst model.}

\item{...}{Currently not used.}
}
\description{
Update BFGrid
}
