% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_curve.R
\name{area_under_curve}
\alias{area_under_curve}
\alias{auc}
\title{Area under the Curve (AUC)}
\usage{
area_under_curve(x, y, method = c("trapezoid", "step", "spline"), ...)

auc(x, y, method = c("trapezoid", "step", "spline"), ...)
}
\arguments{
\item{x}{Vector of x values.}

\item{y}{Vector of y values.}

\item{method}{Can be "trapezoid" (default), "step" or "spline".}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Based on the DescTools \code{AUC} function. It calculates the area under the curve with a naive algorithm and with a more elaborated spline approach. The curve must be given by vectors of xy-coordinates.
}
\details{
If method is set to "trapezoid" then the curve is formed by connecting all points by a direct line (composite trapezoid rule). If "step" is chosen then a stepwise connection of two points is used. For calculating the area under a spline interpolation the splinefun function is used in combination with integrate. The AUC function will handle unsorted x values (by sorting x) and ties for the x values (by ignoring duplicates).
}
\examples{
library(bayestestR)
posterior <- distribution_normal(1000)

dens <- as.data.frame(density(posterior))
dens <- dens[dens$x > 0, ]
x <- dens$x
y <- dens$y

area_under_curve(x, y, method = "trapezoid")
area_under_curve(x, y, method = "step")
area_under_curve(x, y, method = "spline")
}
\seealso{
DescTools
}
