% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{as.data.frame.density}
\alias{as.data.frame.density}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{density}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce to a Data Frame
}
