% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_estimate.R, R/p_direction.R, R/p_map.R,
%   R/p_rope.R
\name{as.numeric.map_estimate}
\alias{as.numeric.map_estimate}
\alias{as.numeric.p_direction}
\alias{as.numeric.p_map}
\alias{as.numeric.p_rope}
\title{Convert to Numeric}
\usage{
\method{as.numeric}{map_estimate}(x, ...)

\method{as.numeric}{p_direction}(x, ...)

\method{as.numeric}{p_map}(x, ...)

\method{as.numeric}{p_rope}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Convert to Numeric
}
