% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor.R
\name{bayesfactor}
\alias{bayesfactor}
\alias{bayesfactor.numeric}
\alias{bayesfactor_savagedickey}
\title{Compute Bayes Factor (BF)}
\usage{
bayesfactor(x, method = "savage-dickey", ...)

\method{bayesfactor}{numeric}(x, method = "savage-dickey", ...)

bayesfactor_savagedickey(x, prior, null = 0, direction = 0)
}
\arguments{
\item{x}{Vector representing a posterior distribution}

\item{method}{The method to be used. Currently only \code{"savage-dickey"} is supported}

\item{...}{Arguments passed to the specific methods}

\item{prior}{Vector representing a prior distribution}

\item{null}{Value to be tested}

\item{direction}{Test type. One of \code{-1} (left tailed), \code{0} (defult; two tailed) or \code{1} (right tailed)}
}
\description{
Compute likelihood ratio between two competing hypotheses
}
\references{
Wagenmakers, E. J., Lodewyckx, T., Kuriyal, H., & Grasman, R. (2010). Bayesian
hypothesis testing for psychologists: A tutorial on the Savage-Dickey method.
Cognitive psychology, 60(3), 158-189.
}
\author{
Mattan S. Ben-Shachar
}
