% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor.R
\name{bayesfactor}
\alias{bayesfactor}
\alias{bayesfactor.numeric}
\alias{bayesfactor_savagedickey}
\title{Bayes Factor (BF)}
\usage{
bayesfactor(x, prior, h0 = 0, direction = 0,
  method = "savage-dickey", ...)

\method{bayesfactor}{numeric}(x, prior, h0 = 0, direction = 0,
  method = "savage-dickey", ...)

bayesfactor_savagedickey(x, prior, h0 = 0, direction = 0)
}
\arguments{
\item{x}{Vector representing a posterior distribution.}

\item{prior}{Vector representing a prior distribution.}

\item{h0}{Value to be tested against (usually \code{0} in the context of null hypothesis testing).}

\item{direction}{Test type. One of \code{-1} (left tailed), \code{0} (defult; two tailed) or \code{1} (right tailed).}

\item{method}{The method to be used. Currently only \code{"savage-dickey"} is supported.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute the Bayes Factor, i.e., the likelihood ratio between two competing hypotheses.
}
\examples{
library(bayestestR)

prior <- rnorm_perfect(1000, mean = 0, sd = 1)
posterior <- rnorm_perfect(1000, mean = .5, sd = .3)

bayesfactor(posterior, prior = prior)

}
\references{
Wagenmakers, E. J., Lodewyckx, T., Kuriyal, H., & Grasman, R. (2010). Bayesian
hypothesis testing for psychologists: A tutorial on the Savage-Dickey method.
Cognitive psychology, 60(3), 158-189.
}
\author{
Mattan S. Ben-Shachar
}
