% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor_savagedickey.R
\name{bayesfactor_savagedickey}
\alias{bayesfactor_savagedickey}
\alias{bayesfactor_savagedickey.numeric}
\title{Savage-Dickey density ratio Bayes factor}
\usage{
bayesfactor_savagedickey(posterior, prior, direction = "two-sided",
  hypothesis = 0)

\method{bayesfactor_savagedickey}{numeric}(posterior, prior,
  direction = "two-sided", hypothesis = 0)
}
\arguments{
\item{posterior}{Vector representing a posterior distribution.}

\item{prior}{Vector representing a prior distribution. If a prior is not provided, will sample from \code{~Cauchy(location = hypothesis, scale = sd(posterior))} (but this should be avoided).}

\item{direction}{Test type. One of \code{0}, \code{"two-sided"} (defult; two tailed),
\code{-1}, \code{"left"} (left tailed), \code{1}, \code{"right"} (right tailed).}

\item{hypothesis}{Value to be tested against (usually \code{0} in the context of null hypothesis testing).}
}
\value{
A data frame of class \code{BFsd} which contains the Bayes factor representing by how
much \emph{less} the null is likely under the posterior compared to the prior (larger than 1
can be interpereted as evidence against the null).
}
\description{
This method computes the ratio between the density of a single value (typically the null)
in two distributions, typically the posterior vs. the prior distributions.
}
\details{
This method is used to examine if the hypothesis value is less or more likely given the observed data.
}
\examples{
library(bayestestR)

prior <- distribution_normal(1000, mean = 0, sd = 1)
posterior <- distribution_normal(1000, mean = .5, sd = .3)

bayesfactor_savagedickey(posterior, prior)
}
\references{
Wagenmakers, E. J., Lodewyckx, T., Kuriyal, H., & Grasman, R. (2010). Bayesian
hypothesis testing for psychologists: A tutorial on the Savage-Dickey method.
Cognitive psychology, 60(3), 158-189.
}
\author{
Mattan S. Ben-Shachar
}
