% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestestR-package.R
\docType{package}
\name{bayestestR-package}
\alias{bayestestR-package}
\alias{bayestestR}
\title{bayestestR: Describing Effects and their Uncertainty, Existence and
Significance within the Bayesian Framework}
\description{
Existing R packages allow users to easily fit a large variety of models
and extract and visualize the posterior draws. However, most of these
packages only return a limited set of indices (e.g., point-estimates and
CIs). \strong{bayestestR} provides a comprehensive and consistent set of
functions to analyze and describe posterior distributions generated by a
variety of models objects, including popular modeling packages such as
\strong{rstanarm}, \strong{brms} or \strong{BayesFactor}.

References:
\itemize{
\item Makowski et al. (2019) \doi{10.21105/joss.01541}
\item Makowski et al. (2019) \doi{10.3389/fpsyg.2019.02767}
}
}
\details{
\code{bayestestR}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://easystats.github.io/bayestestR/}
  \item Report bugs at \url{https://github.com/easystats/bayestestR/issues}
}

}
\author{
\strong{Maintainer}: Dominique Makowski \email{officialeasystats@gmail.com} (\href{https://orcid.org/0000-0001-5375-9967}{ORCID})

Authors:
\itemize{
  \item Daniel Lüdecke \email{d.luedecke@uke.de} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID})
  \item Mattan S. Ben-Shachar \email{matanshm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-4287-4801}{ORCID})
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID})
  \item Micah K. Wilson \email{micah.k.wilson@curtin.edu.au} (\href{https://orcid.org/0000-0003-4143-7308}{ORCID})
  \item Brenton M. Wiernik \email{brenton@wiernik.org} (\href{https://orcid.org/0000-0001-9560-6336}{ORCID})
}

Other contributors:
\itemize{
  \item Paul-Christian Bürkner \email{paul.buerkner@gmail.com} [reviewer]
  \item Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID}) [reviewer]
  \item Henrik Singmann \email{singmann@gmail.com} (\href{https://orcid.org/0000-0002-4842-3657}{ORCID}) [contributor]
  \item Quentin F. Gronau (\href{https://orcid.org/0000-0001-5510-6943}{ORCID}) [contributor]
  \item Sam Crawley \email{sam@crawley.nz} (\href{https://orcid.org/0000-0002-7847-0411}{ORCID}) [contributor]
}

}
\keyword{internal}
