% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{ci.numeric}
\title{Confidence/Credible Interval}
\usage{
ci(model, ci = 0.9, ...)

\method{ci}{numeric}(model, ci = 0.9, verbose = TRUE, ...)
}
\arguments{
\item{model}{A model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute Confidence/Credible Intervals (CI) for Bayesian and frequentist models.
Documentation is accessible for:
\itemize{
  \item Bayesian models
  \item LM and GLMs
  \item Mixed models
}
}
\examples{
library(bayestestR)

ci(rnorm(1000))

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
ci(model)
ci(model, ci = c(.80, .90, .95))

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
ci(model)
ci(model, ci = c(.80, .90, .95))
}


}
