% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{ci.numeric}
\alias{ci.data.frame}
\alias{ci.emmGrid}
\alias{ci.stanreg}
\alias{ci.brmsfit}
\alias{ci.BFBayesFactor}
\title{Confidence/Credible Interval (CI)}
\usage{
ci(x, ...)

\method{ci}{numeric}(x, ci = 0.89, verbose = TRUE, ...)

\method{ci}{data.frame}(x, ci = 0.89, verbose = TRUE, ...)

\method{ci}{emmGrid}(x, ci = 0.89, verbose = TRUE, ...)

\method{ci}{stanreg}(x, ci = 0.89, effects = c("fixed", "random",
  "all"), parameters = NULL, verbose = TRUE, ...)

\method{ci}{brmsfit}(x, ci = 0.89, effects = c("fixed", "random",
  "all"), component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL, verbose = TRUE, ...)

\method{ci}{BFBayesFactor}(x, ci = 0.89, verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{stanreg} or \code{brmsfit} model, or a vector representing a posterior distribution.}

\item{...}{Currently not used.}

\item{ci}{Value or vector of probability of the CI (between 0 and 1)
to be estimated. Default to \code{.89} (89\%) for Bayesian models and \code{.95} (95\%) for frequentist models.}

\item{verbose}{Toggle off warnings.}

\item{effects}{Should results for fixed effects, random effects or both be returned?
Only applies to mixed models. May be abbreviated.}

\item{parameters}{Regular expression pattern that describes the parameters that
should be returned. Meta-parameters (like \code{lp__} or \code{prior_}) are
filtered by default, so only parameters that typically appear in the
\code{summary()} are returned. Use \code{parameters} to select specific parameters
for the output.}

\item{component}{Should results for all parameters, parameters for the conditional model
or the zero-inflated part of the model be returned? May be abbreviated. Only
applies to \pkg{brms}-models.}
}
\value{
A data frame with following columns:
  \itemize{
    \item \code{Parameter} The model parameter(s), if \code{x} is a model-object. If \code{x} is a vector, this column is missing.
    \item \code{CI} The probability of the credible interval.
    \item \code{CI_low}, \code{CI_high} The lower and upper credible interval limits for the parameters.
  }
}
\description{
Compute Confidence/Credible Intervals (CI) for Bayesian and frequentist models. The Documentation is accessible for:
}
\details{
\itemize{
 \item \href{https://easystats.github.io/bayestestR/articles/credible_interval.html}{Bayesian models}
 \item \href{https://easystats.github.io/parameters/reference/ci.merMod.html}{Frequentist models}
}
}
\examples{
library(bayestestR)

posterior <- rnorm(1000)
ci(posterior)
ci(posterior, ci = c(.80, .89, .95))

df <- data.frame(replicate(4, rnorm(100)))
ci(df)
ci(df, ci = c(.80, .89, .95))

library(rstanarm)
model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
ci(model)
ci(model, ci = c(.80, .89, .95))

library(emmeans)
ci(emtrends(model, ~1, "wt"))
\dontrun{
library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
ci(model)
ci(model, ci = c(.80, .89, .95))

library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
ci(bf)
ci(bf, ci = c(.80, .89, .95))
}

}
