% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\alias{ci.numeric}
\alias{ci.stanreg}
\alias{ci.brmsfit}
\title{Confidence/Credible Interval}
\usage{
ci(posterior, ci = 0.9, ...)

\method{ci}{numeric}(posterior, ci = 0.9, verbose = TRUE, ...)

\method{ci}{stanreg}(posterior, ci = 0.9, effects = c("fixed",
  "random", "all"), parameters = NULL, verbose = TRUE, ...)

\method{ci}{brmsfit}(posterior, ci = 0.9, effects = c("fixed",
  "random", "all"), component = c("conditional", "zi", "zero_inflated",
  "all"), parameters = NULL, verbose = TRUE, ...)
}
\arguments{
\item{posterior}{Vector representing a posterior distribution. Can also be a \code{stanreg} or \code{brmsfit} model.}

\item{ci}{Value or vector of HDI probability (between 0 and 1) to be estimated. Named Credible Interval (CI) for consistency.}

\item{...}{Currently not used.}

\item{verbose}{Toggle off warnings.}

\item{effects}{Should results for fixed effects, random effects or both be returned?
Only applies to mixed models. May be abbreviated.}

\item{parameters}{Regular expression pattern that describes the parameters that
should be returned. Meta-parameters (like \code{lp__} or \code{prior_}) are
filtered by default, so only parameters that typically appear in the
\code{summary()} are returned. Use \code{parameters} to select specific parameters
for the output.}

\item{component}{Should results for all parameters, parameters for the conditional model
or the zero-inflated part of the modelbe returned? May be abbreviated. Only
applies to \pkg{brms}-models.}
}
\description{
Compute Confidence/Credible Intervals (CI) for Bayesian and frequentist models.
Documentation is accessible for:
\itemize{
  \item Bayesian models
  \item LM and GLMs
  \item Mixed models
}
}
\examples{
library(bayestestR)

ci(rnorm(1000))

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
ci(model)
ci(model, ci = c(.80, .90, .95))

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
ci(model)
ci(model, ci = c(.80, .90, .95))}

}
