% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{density_at}
\alias{density_at}
\alias{probability_at}
\title{Probability of a Given Point}
\usage{
density_at(posterior, x, precision = 2^10, ...)

probability_at(posterior, x, precision = 2^10, ...)
}
\arguments{
\item{posterior}{Vector representing a posterior distribution.}

\item{x}{The value of which to get the approximate probability.}

\item{precision}{Number of points of density data. See the \code{n} parameter in \link[=density]{density}.}

\item{...}{further arguments for (non-default) methods.}
}
\description{
Compute the density of a given point of a distribution.
}
\examples{
library(bayestestR)
posterior <- distribution_normal(n = 10)
density_at(posterior, 0)
density_at(posterior, c(0, 1))
}
