% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_posterior.R
\name{describe_posterior}
\alias{describe_posterior}
\alias{describe_posterior.stanreg}
\alias{describe_posterior.brmsfit}
\title{Describe Posterior Distributions}
\usage{
describe_posterior(posteriors, estimate = "median", ci = 0.9,
  ci_method = "hdi", test = c("pd", "rope"), rope_range = "default",
  rope_full = TRUE, dispersion = TRUE, ...)

\method{describe_posterior}{stanreg}(posteriors, estimate = "median",
  ci = 0.9, ci_method = "hdi", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE,
  effects = c("fixed", "random", "all"), parameters = NULL, ...)

\method{describe_posterior}{brmsfit}(posteriors, estimate = "median",
  ci = 0.9, ci_method = "hdi", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE,
  effects = c("fixed", "random", "all"), component = c("conditional",
  "zi", "zero_inflated", "all"), parameters = NULL, ...)
}
\arguments{
\item{posteriors}{A vector, dataframe or model of posterior draws.}

\item{estimate}{The \href{https://easystats.github.io/bayestestR/articles/indicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{ci}{Value or vector of probability of the interval (between 0 and 1)
to be estimated. Named Credible Interval (CI) for consistency.}

\item{ci_method}{The type of index used for Credible Interval. Can be \link{hdi} (default) or "quantile" (see \link{ci}).}

\item{test}{The \href{https://easystats.github.io/bayestestR/articles/indicesEstimationComparison.html}{indices of effect existence} to compute. Can be a character or a list with "p_direction", "rope", "p_map" or "bayesfactor".}

\item{rope_range}{\href{https://easystats.github.io/bayestestR/#rope}{ROPE's} lower and higher bounds. Should be a list of two values (e.g., \code{c(-0.1, 0.1)}) or \code{"default"}. If \code{"default"}, the bounds are set to \code{x +- 0.1*SD(response)}.}

\item{rope_full}{If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the \code{ci} argument.}

\item{dispersion}{if \code{TRUE}, computes indices of dispersion related to the estimate(s) (\code{SD} and \code{MAD} for \code{mean} and \code{median}, respectively).}

\item{...}{Additional arguments to be passed to or from methods.}

\item{effects}{Should results for fixed effects, random effects or both be returned?
Only applies to mixed models. May be abbreviated.}

\item{parameters}{Regular expression pattern that describes the parameters that
should be returned. Meta-parameters (like \code{lp__} or \code{prior_}) are
filtered by default, so only parameters that typically appear in the
\code{summary()} are returned. Use \code{parameters} to select specific parameters
for the output.}

\item{component}{Should results for all parameters, parameters for the conditional model
or the zero-inflated part of the model be returned? May be abbreviated. Only
applies to \pkg{brms}-models.}
}
\description{
Compute indices relevant to describe and characterise the posterior distributions.
}
\examples{
describe_posterior(rnorm(1000))

df <- data.frame(replicate(4, rnorm(100)))
describe_posterior(df, estimate = "all", dispersion = TRUE)
describe_posterior(df, estimate = c("median", "MAP"))

}
