% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_posterior.R
\name{describe_posterior}
\alias{describe_posterior}
\alias{describe_posterior.numeric}
\alias{describe_posterior.double}
\alias{describe_posterior.data.frame}
\title{Describe posterior distributions.}
\usage{
describe_posterior(posteriors, ci = 0.9, ci_method = "hdi",
  estimate = "median", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE)

\method{describe_posterior}{numeric}(posteriors, ci = 0.9,
  ci_method = "hdi", estimate = "median", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE)

\method{describe_posterior}{double}(posteriors, ci = 0.9,
  ci_method = "hdi", estimate = "median", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE)

\method{describe_posterior}{data.frame}(posteriors, ci = 0.9,
  ci_method = "hdi", estimate = "median", test = c("pd", "rope"),
  rope_range = "default", rope_full = TRUE, dispersion = TRUE)
}
\arguments{
\item{posteriors}{A vector, dataframe or model of posterior draws.}

\item{ci}{Credible Interval (CI) level. Default to 0.90 (90\%).}

\item{ci_method}{The type of index used for Credible Interval. Can be \link{hdi} (default) or "quantile" (see \link{ci}).}

\item{estimate}{The \href{https://easystats.github.io/bayestestR/articles/indicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{test}{The \href{https://easystats.github.io/bayestestR/articles/indicesEstimationComparison.html}{indices of effect existence} to compute. Can be a character or a list with "p_direction", "rope", "p_map" or "bayesfactor".}

\item{rope_range}{\href{https://easystats.github.io/bayestestR/#rope}{ROPE's} lower and higher bounds. Should be a list of two values (e.g., \code{c(-0.1, 0.1)}) or \code{"default"}. If \code{"default"}, the bounds are set to \code{x +- 0.1*SD(response)}.}

\item{rope_full}{If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the \code{ci} argument.}

\item{dispersion}{Indices of dispersion related to the estimate(s) (\code{SD} and \code{MAD} for \code{mean} and \code{median}, respectively).}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
Compute indices relevant to describe and characterise the posterior distributions.
}
\examples{
describe_posterior(rnorm(1000))


}
