% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_distribution.R
\name{distribution}
\alias{distribution}
\alias{distribution_normal}
\alias{rnorm_perfect}
\alias{rcauchy_perfect}
\alias{rpois_perfect}
\alias{rt_perfect}
\title{Empirical Distributions}
\usage{
distribution(type = "normal", ...)

distribution_normal(n, mean = 0, sd = 1, random = FALSE)

rnorm_perfect(n, mean = 0, sd = 1)

rcauchy_perfect(n, location = 0, scale = 1)

rpois_perfect(n, lambda)

rt_perfect(n, df, ncp)
}
\arguments{
\item{type}{Can be \code{"normal"} (default), \code{"cauchy"}, \code{"poisson"} or \code{"student"}.}

\item{...}{Arguments passed to or from other methods.}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{random}{Generate near-perfect or random (simple wrappers for the base R \code{r*} functions) distributions.}

\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}

\item{lambda}{vector of (non-negative) means.}

\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer).  \code{df
      = Inf} is allowed.}

\item{ncp}{non-centrality parameter \eqn{\delta}{delta};
    currently except for \code{rt()}, only for \code{abs(ncp) <= 37.62}.
    If omitted, use the central t distribution.}
}
\description{
Generate a sample of size \code{n} with a near-perfect distribution.
}
\examples{
library(bayestestR)
x <- distribution(n = 10)
plot(density(x))
}
