% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample.R
\name{effective_sample}
\alias{effective_sample}
\alias{effective_sample.brmsfit}
\alias{effective_sample.stanreg}
\title{Effective Sample Size (ESS)}
\usage{
effective_sample(model, ...)

\method{effective_sample}{brmsfit}(
  model,
  effects = c("fixed", "random", "all"),
  component = c("conditional", "zi", "zero_inflated", "all"),
  parameters = NULL,
  ...
)

\method{effective_sample}{stanreg}(
  model,
  effects = c("fixed", "random", "all"),
  component = c("location", "all", "conditional", "smooth_terms", "sigma",
    "distributional", "auxiliary"),
  parameters = NULL,
  ...
)
}
\arguments{
\item{model}{A \code{stanreg}, \code{stanfit}, \code{brmsfit}, \code{blavaan}, or \code{MCMCglmm} object.}

\item{...}{Currently not used.}

\item{effects}{Should results for fixed effects, random effects or both be
returned? Only applies to mixed models. May be abbreviated.}

\item{component}{Should results for all parameters, parameters for the
conditional model or the zero-inflated part of the model be returned? May
be abbreviated. Only applies to \pkg{brms}-models.}

\item{parameters}{Regular expression pattern that describes the parameters
that should be returned. Meta-parameters (like \code{lp__} or \code{prior_}) are
filtered by default, so only parameters that typically appear in the
\code{summary()} are returned. Use \code{parameters} to select specific parameters
for the output.}
}
\value{
A data frame with two columns: Parameter name and effective sample size (ESS).
}
\description{
Effective Sample Size (ESS) is a measure of how much independent information
there is in autocorrelated chains. It is used to assess the quality of MCMC
samples. A higher ESS indicates more reliable estimates. For most
applications, an effective sample size greater than 1,000 is sufficient for
stable estimates (B端rkner, 2017). This function returns the effective sample
size (ESS) for various Bayesian model objects. For \code{brmsfit} objects, the
returned ESS corresponds to the bulk-ESS (and the tail-ESS is also returned).
}
\details{
\itemize{
\item \strong{Effective Sample (ESS)} should be as large as possible, altough
for most applications, an effective sample size greater than 1,000 is
sufficient for stable estimates (B端rkner, 2017). The ESS corresponds to the
number of independent samples with the same estimation power as the N
autocorrelated samples. It is is a measure of \dQuote{how much independent
information there is in autocorrelated chains} (\emph{Kruschke 2015, p182-3}).
\item \strong{Bulk-ESS} is useful as a diagnostic for the sampling efficiency in
the bulk of the posterior. It is defined as the effective sample size for
rank normalized values using split chains. It can be interpreted as the
reliability of indices of central tendency (mean, median, etc.).
\item \strong{Tail-ESS} is useful as a diagnostic for the sampling efficiency in
the tails of the posterior. It is defined as the minimum of the effective
sample sizes for 5\% and 95\% quantiles. It can be interpreted as the
reliability of indices that depend on the tails of the distribution (e.g.,
credible intervals, tail probabilities, etc.).
}
}
\examples{
\dontshow{if (all(insight::check_if_installed(c("rstanarm", "brms"), quietly = TRUE))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
model <- suppressWarnings(rstanarm::stan_glm(
  mpg ~ wt + gear,
  data = mtcars,
  chains = 2,
  iter = 200,
  refresh = 0
))
effective_sample(model)

model <- brms::brm(
  mpg ~ wt,
  data = mtcars,
  chains = 2,
  iter = 200,
  refresh = 0
)
effective_sample(model)
}
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Kruschke, J. (2014). Doing Bayesian data analysis: A tutorial with R, JAGS,
and Stan. Academic Press.
\item B端rkner, P. C. (2017). brms: An R package for Bayesian multilevel models
using Stan. Journal of Statistical Software, 80(1), 1-28
\item Vehtari, A., Gelman, A., Simpson, D., Carpenter, B., & B端rkner, P.-C.
(2021). Rank-normalization, folding, and localization: An improved R-hat
for assessing convergence of MCMC. Bayesian Analysis, 16(2), 667-718.
}
}
