% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_density.R
\name{estimate_density}
\alias{estimate_density}
\title{Density Estimation}
\usage{
estimate_density(x, method = "kernel", precision = 2^10, bw = "SJ",
  ...)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.  For the
    default method a numeric vector: long vectors are not supported.}

\item{method}{Method of density estimation.}

\item{precision}{Number of points of density data. See the \code{n} parameter in \link[=density]{density}.}

\item{bw}{the smoothing bandwidth to be used.  The kernels are scaled
    such that this is the standard deviation of the smoothing kernel.
    (Note this differs from the reference books cited below, and from S-PLUS.)

    \code{bw} can also be a character string giving a rule to choose the
    bandwidth.  See \code{\link{bw.nrd}}. \cr The default,
    \code{"nrd0"}, has remained the default for historical and
    compatibility reasons, rather than as a general recommendation,
    where e.g., \code{"SJ"} would rather fit, see also Venables and
    Ripley (2002).

    The specified (or computed) value of \code{bw} is multiplied by
    \code{adjust}.
  }

\item{...}{further arguments for (non-default) methods.}
}
\description{
Density Estimation
}
\examples{
x <- rnorm(250, 1)

density_kernel <- estimate_density(x, method = "kernel")
density_logspline <- estimate_density(x, method = "logspline")
density_KernSmooth <- estimate_density(x, method = "KernSmooth")

hist(x, prob = TRUE)
lines(density_kernel$x, density_kernel$y, col = "black", lwd = 2)
lines(density_logspline$x, density_logspline$y, col = "red", lwd = 2)
lines(density_KernSmooth$x, density_KernSmooth$y, col = "blue", lwd = 2)
}
