% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flatten_list.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a list}
\usage{
flatten_list(object, name = "name")
}
\arguments{
\item{object}{A list.}

\item{name}{Name of column of keys in the case the output is a dataframe.}
}
\description{
Flatten a list
}
