% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{hdi}
\alias{hdi}
\title{Highest Density Interval (HDI)}
\usage{
hdi(posterior, prob = 0.9, verbose = TRUE)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{prob}{value or vector between 0 and 1, indicating the interval that is to be estimated.}

\item{verbose}{toggle off warnings.}
}
\description{
Compute the Highest Density Interval (HDI) of a posterior distribution, i.e., the interval which contains all points within the interval have a higher probability density than points outside the interval.
}
\details{
By default, hdi() returns the 90% intervals, deemed to be more stable than, for instance, 95% intervals (Kruschke 2015).
}
\examples{
library(bayestestR)

posterior <- rnorm(1000)
hdi(posterior, prob = 0.9)
hdi(posterior, prob = c(0.8, 0.9, 0.95))
}
\author{
All credits go to \href{https://rdrr.io/cran/ggdistribute/src/R/stats.R}{ggdistribute}.
}
