% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_estimate.R
\name{map_estimate}
\alias{map_estimate}
\title{Maximum A Posteriori (MAP) Estimate}
\usage{
map_estimate(posterior, precision = 2^10)
}
\arguments{
\item{posterior}{Vector representing a posterior distribution. Can also be a \code{stanreg} or \code{brmsfit} model.}

\item{precision}{Number of points for density estimation. See the \code{n} parameter in \link[=density]{density}.}
}
\value{
The x, y coordinates of the MAP, corresponding to the MAP (x) and its density (y).
}
\description{
Find the Highest Maximum A Posteriori (MAP) estimate of a posterior. It corresponds to the 'peak' of the posterior distribution.
}
\examples{
library(bayestestR)

posterior <- rnorm(1000)
map_estimate(posterior)

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
map_estimate(model)

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
map_estimate(model)
}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
