% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse.R
\name{mcse}
\alias{mcse}
\alias{mcse.stanreg}
\title{Compute Monte-Carlo Standard Error}
\usage{
mcse(model, ...)

\method{mcse}{stanreg}(model, effects = c("fixed", "random", "all"),
  parameters = NULL, ...)
}
\arguments{
\item{model}{A \code{stanreg}, \code{stanfit}, or \code{brmsfit} object.}

\item{...}{Not used}

\item{effects}{Should results for fixed effects, random effects or both be returned?
Only applies to mixed models. May be abbreviated.}

\item{parameters}{Regular expression pattern that describes the parameters that
should be returned. Meta-parameters (like \code{lp__} or \code{prior_}) are
filtered by default, so only parameters that typically appear in the
\code{summary()} are returned. Use \code{parameters} to select specific parameters
for the output.}
}
\value{
\code{mcse()} and \code{n_eff()} return a tibble with two columns: one
  with the term names and one with the related statistic resp. effective
  sample size.
}
\description{
\code{mcse()} returns the Monte Carlo standard error.
}
\details{
The Monte Carlo Standard Error is another useful measure of accuracy of
 the chains. It is defined as standard deviation of the chains divided by
 their effective sample size (the formula for \code{mcse()} is from
 Kruschke 2015, p. 187). The MCSE \dQuote{provides a quantitative suggestion
 of how big the estimation noise is}.
}
\examples{
\dontrun{
if (require("rstanarm")) {
  m <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
  mcse(m)
}}
}
\references{
Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
}
