% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\alias{mediation.brmsfit}
\alias{mediation.stanmvreg}
\title{Summary of Bayesian multivariate-response mediation-models}
\usage{
mediation(model, ...)

\method{mediation}{brmsfit}(
  model,
  treatment,
  mediator,
  response = NULL,
  centrality = "median",
  ci = 0.89,
  method = "ETI",
  ...
)

\method{mediation}{stanmvreg}(
  model,
  treatment,
  mediator,
  response = NULL,
  centrality = "median",
  ci = 0.89,
  method = "ETI",
  ...
)
}
\arguments{
\item{model}{A \code{brmsfit} object.}

\item{...}{Not used.}

\item{treatment}{Character, name of the treatment variable (or direct effect)
in a (multivariate response) mediator-model. If missing, \code{mediation()}
tries to find the treatment variable automatically, however, this may fail.}

\item{mediator}{Character, name of the mediator variable in a (multivariate
response) mediator-model. If missing, \code{mediation()} tries to find the
treatment variable automatically, however, this may fail.}

\item{response}{A named character vector, indicating the names of the response
variables to be used for the mediation analysis. Usually can be \code{NULL},
in which case these variables are retrieved automatically. If not \code{NULL},
names should match the names of the model formulas (\code{names(model$formula$forms)}).}

\item{centrality}{The point-estimates (centrality indices) to compute.  Character (vector) or list with one or more of these options: \code{"median"}, \code{"mean"}, \code{"MAP"} or \code{"all"}.}

\item{ci}{Value or vector of probability of the CI (between 0 and 1)
to be estimated. Default to \code{.89} (89\%) for Bayesian models and \code{.95} (95\%) for frequentist models.}

\item{method}{Can be \link[=eti]{'ETI'} (default), \link[=hdi]{'HDI'} or \link[=si]{'SI'}.}
}
\value{
A data frame with direct, indirect, mediator and
  total effect of a multivariate-response mediation-model, as well as the
  proportion mediated. The effect sizes are median values of the posterior
  samples (use \code{centrality} for other centrality indices).
}
\description{
\code{mediation()} is a short summary for multivariate-response
  mediation-models.
}
\details{
\code{mediation()} returns a data frame with information on the
      \emph{direct effect} (mean value of posterior samples from \code{treatment}
      of the outcome model), \emph{mediator effect} (mean value of posterior
      samples from \code{mediator} of the outcome model), \emph{indirect effect}
      (mean value of the multiplication of the posterior samples from
      \code{mediator} of the outcome model and the posterior samples from
      \code{treatment} of the mediation model) and the total effect (mean
      value of sums of posterior samples used for the direct and indirect
      effect). The \emph{proportion mediated} is the indirect effect divided
      by the total effect.
      \cr \cr
      For all values, the 89\% credible intervals are calculated by default.
      Use \code{ci} to calculate a different interval.
      \cr \cr
      The arguments \code{treatment} and \code{mediator} do not necessarily
      need to be specified. If missing, \code{mediation()} tries to find the
      treatment and mediator variable automatically. If this does not work,
      specify these variables.
      \cr \cr
      The direct effect is also called \emph{average direct effect} (ADE),
      the indirect effect is also called \emph{average causal mediation effects}
      (ACME).
}
\seealso{
The \pkg{mediation} package for a causal mediation analysis in
  the frequentist framework.
}
