% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_direction.R
\name{p_direction}
\alias{p_direction}
\title{Probability of Direction (pd)}
\usage{
p_direction(posterior)
}
\arguments{
\item{posterior}{Vector representing a posterior distribution. Can also be a \code{stanreg} or \code{brmsfit} model.}
}
\description{
Compute the Probability of Direction (pd, also known as the Maximum Probability of Effect - MPE). It varies between 50\% and 100\% and can be interpreted as the probability (expressed in percentage) that a parameter (described by its posterior distribution) is strictly positive or negative (consistently with the median's sign). It is defined as the proportion of the posterior distribution that is of the median's sign. Altough differently expressed, this index is fairly similar to the frequentist p-value (i.e., is strongly correlated).
}
\examples{
library(bayestestR)

# Simulate a posterior distribution of mean 1 and SD 1
posterior <- rnorm(1000, mean = 1, sd = 1)
p_direction(posterior)

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
p_direction(model)

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
p_direction(model)
}
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
