% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rope.R
\name{p_rope}
\alias{p_rope}
\title{ROPE-based p-value}
\usage{
p_rope(posterior, bounds = c(-0.1, 0.1), precision = 0.1)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{bounds}{ROPE's lower and higher bounds.}

\item{precision}{The precision by which to explore the ROPE space. Lower values increase the precision of the returned p value but can be quite computationnaly costly.}
}
\description{
The ROPE-based p-value represents the maximum Credible Interval (\link[=hdi]{HDI}) that does not contain (positive values) or is entirely contained (negative values) in the negligible values space defined by the \link[=rope]{ROPE}. A ROPE-based p of 97% means that there is a probability of .97 that a parameter (desccribed by its posterior distribution) is outside the ROPE.
}
\examples{
library(bayestestR)

p_rope(posterior=rnorm(1000, mean = 1, sd = 1), bounds = c(-0.1, 0.1))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
