% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rope.R
\name{p_rope}
\alias{p_rope}
\title{ROPE-based p-value}
\usage{
p_rope(posterior, bounds = c(-0.1, 0.1), precision = 0.1)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{bounds}{ROPE's lower and higher bounds.}

\item{precision}{The precision by which to explore the ROPE space. Lower values increase the precision of the returned p value but can be quite computationnaly costly.}
}
\description{
The ROPE-based p-value represents the maximum percentage of \link[=hdi]{HDI} that does not contain (positive values) or is entirely contained (negative values) in the negligible values space defined by the \link[=rope]{ROPE}. It differs from the ROPE, i.e., the proportion of a given CI in the ROPE, by representing the maximum CI to reach a ROPE proportion of 0\% (positive values) or 100\% (negative values). A ROPE-based p of 97\% means that there is a probability of .97 that a parameter (desccribed by its posterior distribution) is outside the ROPE. On the contrary, a ROPE-based p of -97\% means that there is also a probability of 0.97 that the parameter is inside the ROPE.
}
\examples{
library(bayestestR)

p_rope(posterior = rnorm(1000, mean = 1, sd = 1), bounds = c(-0.1, 0.1))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
