% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_to_bf.R
\name{p_to_bf}
\alias{p_to_bf}
\title{Convert p-values to Bayes Factors}
\usage{
p_to_bf(x, n_obs = NULL, log = FALSE)
}
\arguments{
\item{x}{A (frequentist) model object, or a (numeric) vector of p-values.}

\item{n_obs}{Number of observations. When \code{x} is a model object, this is
automatically extracted from the model via \code{insight::n_obs()}.}

\item{log}{If \code{TRUE}, return the \code{log(BF)}.}
}
\value{
The Bayes Factors corresponding to the p-values.
}
\description{
Convert p-values to Bayes Factors
}
\examples{
if (requireNamespace("parameters", quietly = TRUE)) {
  data(iris)
  model <- lm(Petal.Length ~ Sepal.Length + Species, data = iris)
  parameters::p_value(model)
  p_to_bf(model)
}
}
\references{
Wagenmakers, E.-J. (2022). Approximate objective Bayes factors from p-values and sample size: The 3p(sqrt(n)) rule. Preprint available on ArXiv: https://psyarxiv.com/egydq
}
