% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_to_bf.R
\name{p_to_bf}
\alias{p_to_bf}
\alias{p_to_bf.numeric}
\alias{p_to_bf.default}
\title{Convert p-values to Bayes Factors}
\usage{
p_to_bf(x, log = FALSE, ...)

\method{p_to_bf}{numeric}(x, log = FALSE, n_obs = NULL, ...)

\method{p_to_bf}{default}(x, log = FALSE, ...)
}
\arguments{
\item{x}{A (frequentist) model object, or a (numeric) vector of p-values.}

\item{n_obs}{Number of observations. Either length 1, or same length as \code{p}.}
}
\value{
A data frame with the p-values and Bayes factors (against the null).
}
\description{
Convert p-values to Bayes Factors
}
\examples{
if (requireNamespace("parameters", quietly = TRUE)) {
  data(iris)
  model <- lm(Petal.Length ~ Sepal.Length + Species, data = iris)
  p_to_bf(model)
}

}
\references{
\itemize{
\item Wagenmakers, E.J. (2022). Approximate objective Bayes factors from p-values
and sample size: The 3p(sqrt(n)) rule. Preprint available on ArXiv:
https://psyarxiv.com/egydq
}
}
\seealso{
\code{\link[=bic_to_bf]{bic_to_bf()}}
}
