% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rnorm_perfect.R
\name{rnorm_perfect}
\alias{rnorm_perfect}
\alias{rcauchy_perfect}
\alias{rpois_perfect}
\title{Perfect Empirical Distributions}
\usage{
rnorm_perfect(n, mean = 0, sd = 1)

rcauchy_perfect(n, location = 0, scale = 1)

rpois_perfect(n, lambda)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{location}{location and scale parameters.}

\item{scale}{location and scale parameters.}

\item{lambda}{vector of (non-negative) means.}
}
\description{
Generate a sample of size \code{n} with a near-perfect distribution.
}
\examples{
library(bayestestR)
x <- rnorm_perfect(n = 10)
plot(density(x))
}
