% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_perfect.R
\name{rnorm_perfect}
\alias{rnorm_perfect}
\title{The (Perfect) Normal Distribution}
\usage{
rnorm_perfect(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}
}
\description{
Generates a sample of size n with a near-perfect normal distribution.
}
\examples{
library(bayestestR)
x <- rnorm_perfect(n = 10)
plot(density(x))

}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
