% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{rope}
\alias{rope}
\title{Region of Practical Equivalence (ROPE)}
\usage{
rope(posterior, rope = c(-0.1, 0.1), CI = 95)
}
\arguments{
\item{posterior}{vector representing a posterior distribution.}

\item{rope}{ROPE's lower and higher bounds.}

\item{CI}{the credible interval to use.}
}
\description{
Compute the proportion of the HDI of a posterior distribution that lies within a region of practical equivalence.
}
\details{
Statistically, the probability of a posterior distribution of being different from 0 does not make much sense (the probability of it being different from a single point being infinite). Therefore, the idea underlining ROPE is to let the user define an area around the null value enclosing values that are equivalent to the null value for practical purposes (2010, 2011, 2014). Kruschke (2018) suggests that such null value could be set, by default, to the -0.1 to 0.1 range of a standardized parameter (negligible effect size according to Cohen). This could be generalized: For instance, for linear models, the ROPE could be set as 0 +/- .1 * sd(y). Kruschke (2010, 2011, 2014) suggest using the proportion of the 95\% \link[=hdi]{HDI} that falls within the ROPE as an index for "null-hypothesis" testing (as understood under the Bayesian framework, see \link[=rope_test]{rope_test}). Besides the ROPE-based decisions criteria, the proportion of the 95\% CI that falls in the ROPE can be used as a continuous index, which low values are desirable.
}
\examples{
library(bayestestR)

rope(posterior = rnorm(1000, 0, 0.01), rope = c(-0.1, 0.1))
rope(posterior = rnorm(1000, 0, 1), rope = c(-0.1, 0.1))
rope(posterior = rnorm(1000, 1, 0.01), rope = c(-0.1, 0.1))
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
