% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_range.R
\name{rope_range}
\alias{rope_range}
\title{Find Default Equivalence (ROPE) Region Bounds}
\usage{
rope_range(x)
}
\arguments{
\item{x}{Vector representing a posterior distribution. Can also be a \code{stanreg} or \code{brmsfit} model.}
}
\description{
This function attempts at finding suitable "default" values for the Region Of Practical Equivalence (ROPE). Kruschke (2018) suggests that such null value could be set, by default, to the \code{-0.1} to \code{0.1} range of a standardized parameter (negligible effect size according to Cohen, 1988), which can be generalised for linear models to \eqn{[-0.1*SD_{y}, 0.1*SD_{y}]}.
}
\examples{
\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(vs ~ mpg, data = mtcars, family = "binomial")
rope_range(model)

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
rope_range(model)
}

}
\references{
Kruschke, J. K. (2018). Rejecting or accepting parameter values in Bayesian estimation. Advances in Methods and Practices in Psychological Science, 1(2), 270-280. \doi{10.1177/2515245918771304}.
}
