% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_range.R
\name{rope_range}
\alias{rope_range}
\title{Find Default Equivalence (ROPE) Region Bounds}
\usage{
rope_range(x, ...)
}
\arguments{
\item{x}{A \code{stanreg}, \code{brmsfit} or \code{BFBayesFactor} object.}

\item{...}{Currently not used.}
}
\description{
This function attempts at automatically finding suitable "default"
  values for the Region Of Practical Equivalence (ROPE).
}
\details{
\cite{Kruschke (2018)} suggests that the region of practical
  equivalence could be set, by default, to a range from \code{-0.1} to
  \code{0.1} of a standardized parameter (negligible effect size
  according to Cohen, 1988), which can be generalised for linear models
  to \ifelse{html}{\out{-0.1 * SD<sub>y</sub>, 0.1 * SD<sub>y</sub>}}{\eqn{[-0.1*SD_{y}, 0.1*SD_{y}]}}.
  \cr \cr
  For logistic models, the parameters expressed in log odds ratio can be
  converted to standardized difference through the formula
  \ifelse{html}{\out{sqrt(3)/pi}}{\eqn{\sqrt{3}/\pi}}, resulting in a range
  of \code{-0.055} to \code{-0.055}.
  \cr \cr
  For other models with binary outcome, it is strongly recommended to
  manually specify the rope argument. Currently, the same default is applied
  that for logistic models.
  \cr \cr
  For t-tests, the standard deviation of the response is used,
  similarly to linear models (see above).
  \cr \cr
  For correlations, \code{-0.05, 0.05} is used, i.e., half the value of a
  negligible correlation as suggested by Cohen's (1988) rules of thumb.
  \cr \cr
  For all other models, \code{-0.1, 0.1} is used to determine the ROPE limitss,
  but it is strongly advised to specify it manually.
}
\examples{
\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(vs ~ mpg, data = mtcars, family = "binomial")
rope_range(model)

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
rope_range(model)

library(BayesFactor)
bf <- ttestBF(x = rnorm(100, 1, 1))
rope_range(bf)
}

}
\references{
Kruschke, J. K. (2018). Rejecting or accepting parameter values in Bayesian estimation. Advances in Methods and Practices in Psychological Science, 1(2), 270-280. \doi{10.1177/2515245918771304}.
}
