% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_correlation.R
\name{simulate_correlation}
\alias{simulate_correlation}
\title{Data Simulation}
\usage{
simulate_correlation(n = 100, r = 0.5, mean = 0, sd = 1)
}
\arguments{
\item{n}{The number of observations to be generated.}

\item{r}{A value or vector corresponding to the desired correlation coefficients.}
}
\description{
Simulate data with specific characteristics.
}
\examples{

# Correlation --------
data <- simulate_correlation(r = 0.5)
cor.test(data$V1, data$V2)

# Specify mean and SD
data <- simulate_correlation(r = 0.5, n = 50, mean = c(0, 1), sd = c(0.7, 1.7))
cor.test(data$V1, data$V2)
round(c(mean(data$V1), sd(data$V1)), 1)
round(c(mean(data$V2), sd(data$V2)), 1)

# Generate multiple variables
cor_matrix <- matrix(c(1.0, 0.2, 0.4,
                       0.2, 1.0, 0.3,
                       0.4, 0.3, 1.0),
                     nrow = 3)

data <- simulate_correlation(r = cor_matrix)
cor(data)
}
